pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "pillow>=11.3.0",
    "selenium>=4.34.1",
    "webdriver-manager>=4.0.2",
]
